requirements.txt
pandas==1.5.3
python-dotenv==1.0.0
streamlit==1.19.0
streamlit==1.21.0
